refactor: rename log package from syslog to log
The package lived in the log/ directory but was named syslog, so every
call site aliased it. Rename it to match its import path.
